refactor: management service returns more consistent workflow status
