+++ ManagementService/workflow_management_service.json
@@ -8,7 +8,6 @@
             "potentialNext": [
                 "getFunctionEnvironmentVariables",
                 "uploadFunctionEnvironmentVariables",
-                "getWorkflowStatus",
                 "getWorkflowDetails",
                 "undeployWorkflow",
                 "deployWorkflow",
@@ -140,11 +139,6 @@
             "name": "undeployWorkflow",
             "next": [],
             "potentialNext": ["ManagementServiceExit"]
-        },
-        {
-            "name": "getWorkflowStatus",
-            "next": ["ManagementServiceExit"],
-            "potentialNext": []
         },
         {
             "name": "getWorkflowDetails",
